+++ lib/core/luna_ui.dart
@@ -1,6 +1,7 @@
 export 'luna_ui/appbar.dart';
 export 'luna_ui/decoration.dart';
 export 'luna_ui/list_view.dart';
+export 'luna_ui/navigation_bar.dart';
 export 'luna_ui/popup_menu_button.dart';
 export 'luna_ui/sliver_sticky_header.dart';
 export 'luna_ui/snackbar.dart';